% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{compute_gains}
\alias{compute_gains}
\title{Compute Gains}
\usage{
compute_gains(df, start = min(df$date), end = max(df$date))
}
\arguments{
\item{df}{data frame (usually outcome of get_yahoo_data())}

\item{start}{date to start with in format "YYYY-MM-DD". Default to 4 years in the past.}

\item{end}{date to end with in format "YYYY-MM-DD". Default to today.}
}
\value{
data frame
}
\description{
Computes percentage and absolute value gain.
}
\details{
Based on a data frame, given by the get_yahoo_data() function, computes its percentual and absolute gains.
It can contain multiple stocks.
}
