% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{forecast_stock}
\alias{forecast_stock}
\title{Forecast Stock}
\usage{
forecast_stock(df, start = min(df$date), end = max(df$date), code = NULL)
}
\arguments{
\item{df}{data frame (usually outcome of get_yahoo_data())}

\item{start}{date to start with in format "YYYY-MM-DD". Default to 4 years in the past.}

\item{end}{date to end with in format "YYYY-MM-DD". Default to today.}

\item{code}{string Optional. In case there is more than one code in the data, define which
symbol shall be used with the stock symbol as in yahoo finance. Example: "AAPL".}
}
\value{
list
}
\description{
Forecast a single stock in 30 days in the future
}
\details{
A list is returned with the data frame of predictions for each symbol
present, the STLF plots (trend and seasonality), and gains.
}
