% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{get_yahoo_data}
\alias{get_yahoo_data}
\title{Get Yahoo Data}
\usage{
get_yahoo_data(
  code,
  start = lubridate::today() - lubridate::years(4),
  end = lubridate::today()
)
}
\arguments{
\item{code}{string or list of strings with the stock symbol as in yahoo finance. Example: "AAPL".}

\item{start}{date to start with in format "YYYY-MM-DD". Default to 4 years in the past.}

\item{end}{date to end with in format "YYYY-MM-DD". Default to today.}
}
\value{
data frame
}
\description{
Get a data frame with yahoo data using a date frame.
}
