% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{plot_stock_lines}
\alias{plot_stock_lines}
\title{Plot Stock Lines}
\usage{
plot_stock_lines(df)
}
\arguments{
\item{df}{data frame (usually outcome of get_yahoo_data())}
}
\value{
ggplot
}
\description{
Plot the change of a value in a stock (or multiple ones) as lines.
}
