% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{simulate_portfolio_future}
\alias{simulate_portfolio_future}
\title{Simulate Portfolio Future}
\usage{
simulate_portfolio_future(df, investment_df)
}
\arguments{
\item{df}{data frame (usually outcome of get_yahoo_data())}

\item{investment_df}{data frame with columns symbol and investment.}
}
\value{
list
}
\description{
Simulates a Portfolio of Stocks in the future using STLF method.
}
\details{
A list is returned with the data frame of predictions for each symbol
present, the STLF plots (trend and seasonality), gains, and the ggplot of the
multiple stocks in the portfolio.
}
