% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{simulate_portfolio_past}
\alias{simulate_portfolio_past}
\title{Simulate Portfolio Past}
\usage{
simulate_portfolio_past(
  df,
  investment_df,
  start = min(df$date),
  end = max(df$date)
)
}
\arguments{
\item{df}{data frame (usually outcome of get_yahoo_data())}

\item{investment_df}{data frame with columns symbol and investment.}

\item{start}{date to start with in format "YYYY-MM-DD". Default to 4 years in the past.}

\item{end}{date to end with in format "YYYY-MM-DD". Default to today.}
}
\value{
data frame
}
\description{
Simulates the result of a specific investment in given stock(s).
}
\details{
Makes a simulation of investment based only in the past.
}
